+++ src/graphics/functions/linear_algebra.hpp
@@ -7,11 +7,11 @@ namespace gfx {
 
     // Returns a flattened square matrix of dimension n-1 calculated from a matrix with dimension n
     // with the specified row and column removed
-    [[nodiscard]] std::vector<float> getSubmatrix(const std::vector<float>& matrix_values,
+    [[nodiscard]] std::vector<double> getSubmatrix(const std::vector<double>& matrix_values,
                                                   size_t row_to_remove,
                                                   size_t col_to_remove);
 
     // Recursively calculates the determinant of any square matrix from a vector representing
     // their values stored in flattened, row-major order
-    [[nodiscard]] float calculateDeterminant(const std::vector<float>& matrix_values);
+    [[nodiscard]] double calculateDeterminant(const std::vector<double>& matrix_values);
 }
